requirements.txt
fastapi
yandex_music
lastfm-py
uvicorn
